+++ node-test/test1/test.ml
@@ -64,7 +64,7 @@ let uncaught_handler p =
       exit 1
     )
 
-(*** FileSystem **)
+(** FileSystem **)
 
 let root : unit Promise.t =
   let open Promise in
@@ -133,3 +133,15 @@ let () =
   let res2 = Concat.apply Path2.join args in
   assert (res1 = res2);
   ()
+
+(** Arrays **)
+let () =
+    let open Arrays.StringArray in
+    let a = create () in
+    for k = 0 to 10 do
+      push a (string_of_int k);
+    done;
+    let s = join a "," in
+    List.iteri (fun k x ->
+      assert (string_of_int k = x)
+    ) (String.split_on_char ',' s)